requirements.txt
flask
sqlalchemy
psycopg2-binary
pyjwt
werkzeug